Fixed the_ant()
It was just a stupid rounding error that was impossible to diagnose by just debugging, because of my spaghetti code. I managed to work it out by completely starting from scratch for the function on paper whilst on holiday. I also changed the old goround() function to gohead() as it is the same as goooo() but keeping the heading. The new goround() and its counterpart goroundminus() are quick parameter filling shortcuts of gohead() the can fix rounding errors and boundary placement errors.